==> script/02_predict_stability.ipynb <==
import pandas as pd
import numpy as  np
import pickle
import matplotlib as mpl
import matplotlib.pyplot as plt

from functools import partial
from pymatgen.core.periodic_table import Element

from mendeleev import element
from sklearn.model_selection import RandomizedSearchCV, GridSearchCV
from sklearn.model_selection import KFold
from sklearn.ensemble import GradientBoostingRegressor, ExtraTreesRegressor, RandomForestRegressor
from sklearn.metrics import mean_absolute_error, r2_score
from sklearn.preprocessing import MinMaxScaler

## Make fingerprint for ML

def get_spin(el):
    """
    Get number of unpaired electron
    """
    
    block = el.block
    group = el.group
    
    if el.symbol == 'Lu':
        return 0
    elif block == 's':
        return 0 if group == 2 else 1
    elif block == 'p':
        return (group - 12 if group < 16 else 19 - group)
    elif block == 'd':
        return (group - 2 if group < 8 else 12 - group if el.symbol != 'Pt' else 0)
    elif block == 'f':
        return (group - 3 if group < 12 else 17 - group)



df = pd.read_csv('../data/cubic_DP_data_with_cif.csv',index_col=0)
fingerprint_df = pd.DataFrame(columns = ['id'])
fingerprint_df['id'] = df['id']


# Get property
prop_list = []
for i,v in df.iterrows():
    prop_row = []
    for site in ['A_atom_1','A_atom_2','B_atom_1','B_atom_2']:
        el = Element(v[site])
        prop_row.extend([
            el.melting_point,        # Melting Point
            el.boiling_point,        # Boiling Point
            el.X,                    # Electronegativity 
            el.average_ionic_radius, # Ionic radius
            get_spin(el),            # # of Unpaired Electron
            el.common_oxidation_states[0],#int(v[site+'_oxi']),     # Oxidation state
            el.group,                # Group
            el.row,                  # Row
        ])
    prop_list.append(prop_row)

prop_arr = np.array(prop_list).T

for i, v in enumerate(prop_arr):
    fingerprint_df[f'prop_{i+1}'] = v

fingerprint_df

# Make one-hot encoded composition

alkali = ['Li','Na','K','Cs','Rb','Be','Mg','Ca','Sr','Ba']
transition = ['Sc','Ti','V','Cr','Mn','Fe','Co','Ni','Cu','Zn','Y','Zr','Nb','Mo','Ru','Rh','Pd',
             'Ag','Cd','Hf','Ta','W','Re','Os','Ir','Pt','Au','Hg'] # Tc removed: radioactive
main = ['Al','Ga','In','Sn','Tl','Pb','Bi','B','Si','Ge','As','Sb','Te']
rare = ['La','Ce','Pr','Nd','Pm','Sm','Eu','Gd','Tb','Dy','Ho','Er','Tm','Yb','Lu']

cn6, cn12 = {}, {}

for i in alkali+transition+rare+main:
    elem = element(i)
    ionic_radii = elem.ionic_radii
    for radii in ionic_radii:
        if radii.coordination == 'VI':
            for charge in range(1,9):
                if radii.charge == charge:
                    cn6['B'+'_'+i+'_'+str(radii.charge)] = radii.ionic_radius

        elif radii.coordination == 'XII':
            for charge in range(1,9):
                if radii.charge == charge:
                    cn12['A'+'_'+i+'_'+str(radii.charge)] = radii.ionic_radius


elem_candidate = [i for i in cn12.keys()] + [i for i in cn6.keys()] # +1 ~ +7
ohe_list = []

for i,v in df.iterrows():
    dummy = np.zeros(len(elem_candidate))
    elem_lst = ['A_'+v['A_atom_1']+'_'+str(v['A_atom_1_oxi']),'A_'+v['A_atom_2']+'_'+str(v['A_atom_2_oxi']),
               'B_'+v['B_atom_1']+'_'+str(v['B_atom_1_oxi']),'B_'+v['B_atom_2']+'_'+str(v['B_atom_2_oxi'])]
    
    for elem in elem_lst:
        dummy[elem_candidate.index(elem)] += 1
    ohe_list.append(dummy)
    
ohe_list = np.array(ohe_list).T

ohe_dict = {column: fp for column, fp in zip(elem_candidate, ohe_list)}
ohe_df = pd.DataFrame(ohe_dict)

fingerprint_df = pd.concat([fingerprint_df, ohe_df], axis=1)
fingerprint_df

## Train and predict

energy_dict = pd.read_csv('../data/stability_data.csv',index_col=1).to_dict(orient='index')
fingerprint_df['pbx'] = [energy_dict[i]['Pourbaix stability OER'] for i in fingerprint_df['id']]
fingerprint_df['hull'] = [energy_dict[i]['Energy above hull'] for i in fingerprint_df['id']]
fingerprint_df

# PBX stability

calculated_df = fingerprint_df.copy(deep=True).dropna()

X = calculated_df.iloc[:,1:-2].to_numpy()
y = calculated_df.iloc[:,-2].to_numpy()

reg = GradientBoostingRegressor(n_estimators=1321, learning_rate=0.15,subsample=0.6,
                                max_depth=5,random_state=42)

kf = KFold(n_splits=5,shuffle=True,random_state=42)

mae_lst = []
r2_lst = []
for i, (train_idx,test_idx) in enumerate(kf.split(X)):
    X_train,X_test,y_train,y_test = X[train_idx], X[test_idx], y[train_idx], y[test_idx]
    
    scaler = MinMaxScaler()
    X_train = scaler.fit_transform(X_train)
    X_test = scaler.transform(X_test)
    
    reg.fit(X_train,y_train)
    y_pred = reg.predict(X_test)
    
    mae_lst.append(mean_absolute_error(y_test,y_pred))
    r2_lst.append(r2_score(y_test,y_pred))

print(f'MAE : {np.mean(mae_lst)}')
print(f'R2 score : {np.mean(r2_lst)}')

# Energy above hull

X = calculated_df.iloc[:,33:-2].to_numpy()
y = calculated_df.iloc[:,-1].to_numpy()

reg = GradientBoostingRegressor(n_estimators=131, learning_rate=0.28, subsample=0.52,
                               max_depth=19, random_state=42)

kf = KFold(n_splits=5,shuffle=True,random_state=42)

mae_lst = []
r2_lst = []
for i, (train_idx,test_idx) in enumerate(kf.split(X)):
    X_train,X_test,y_train,y_test = X[train_idx], X[test_idx], y[train_idx], y[test_idx]
    
    scaler = MinMaxScaler()
    X_train = scaler.fit_transform(X_train)
    X_test = scaler.transform(X_test)
    
    reg.fit(X_train,y_train)
    y_pred = reg.predict(X_test)
    
    mae_lst.append(mean_absolute_error(y_test,y_pred))
    r2_lst.append(r2_score(y_test,y_pred))

print(f'MAE : {np.mean(mae_lst)}')
print(f'R2 score : {np.mean(r2_lst)}')
