==> script/03_generate_Surface.ipynb <==
import pandas as pd
import numpy as np
import os
import shutil
import tqdm

from ase.io import read, write
from pymatgen.core import Structure
from pymatgen.io.ase import AseAtomsAdaptor

## Calculate cell lattice

df = pd.read_csv('../data/calculated_DP_data_with_cif.csv',index_col=0)
calculated_bulk_list = [Structure.from_str(i,fmt='cif') for i in df['calculated_bulk_cif']]


prototype_surface = read('./CaTiO3_surface.traj')
reference_cell = prototype_surface.cell[0][0]/2

cell_factor_list = []

for structures in tqdm.tqdm(calculated_bulk_list):    
    atoms = AseAtomsAdaptor.get_atoms(structures)
    cell_factor_list.append(atoms.cell[0][0]/reference_cell)
    
df['cell_factor'] = cell_factor_list

## Construct bare structure

order = []
for index, row in df.iterrows():
    B1=row['B_atom_1']
    B2=row['B_atom_2']
    A1=row['A_atom_1']
    A2=row['A_atom_2']
    tmp = [[B1, A1, B2, A2],
           [B1, A2, B2, A1],
           [B2, A1, B1, A2],
           [B2, A2, B1, A1]]
    order.append(tmp)
df['slab_ordering'] = order

tags = ['nn','pn','np','pp']

for index, row in df.iterrows():
    atoms = prototype.copy()
    cell = prototype.cell
    id_ = row['id'].split('_')[1]
    
    new_cell = cell*row['calculated_cell_factor']
    atoms.set_cell(new_cell, scale_atoms=True)
    
    for i, order in enumerate(row['slab_ordering']):
        for atom in atoms:
            if atom.tag == 1:
                atom.symbol = order[0]
            elif atom.tag == 2:
                atom.symbol = order[1]
            elif atom.tag == 3:
                atom.symbol = order[2]
            elif atom.tag == 4:
                atom.symbol = order[3]
                
        # write(dst+ 'SR_' + id_+ 'l'+ tags[i]+'b.traj', atoms)

## Add adsorbate to bare

from ase import Atoms
from ase.io import read,write
# Let's add O*

bare = 'bare surface source'
o = 'O* adsorbed surface destination'
oh = 'OH* adsorbed surface destination'

for file in os.listdir(bare):
    atoms = read(bare+str(file))
    name = file.split('b.traj')[0]
    
    copy_atoms = atoms.copy()
    positions = [i for i in atoms if i.tag == 1][0].position
    copy_atoms += Atoms('O', [positions + np.array([0,0,2])] )
    write(o + name + 'o.traj', copy_atoms)
    
    copy_atoms += Atoms('H', [positions + np.array([0,0.8,2.5])] )
    write(oh + name + 'oh.traj', copy_atoms)
